ENHANCEMENT: several new scripts added
- spw-channel-dumppackagelist
Dumps E:NVR.A strings for every package in a chosen channel to a text file.
Can work recursively and dump all files to an output directory.

- spw-channel-update
Updates a software channel to match one from another satellite, using a package
index and a list of NVR.A strings (in a text file).
Adds only, does not remove. Works when ISS is not available

- spw-errata-details
pretty-prints errata information for the given list of errata to stdout

- spw-package-audit
Compares a software channel to the output of rpm -qa (read from a text file)
reports on newer/older/missing packages in a channel based on the source system.

Also updated packaging MANIFEST template with correct README files
and setup.py to use glob.glob to find script names, rather than explicitly listing them.